+++ app/Exceptions/WordNotFoundException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Exceptions;
+
+use Exception;
+
+class WordNotFoundException extends Exception
+{
+    //
+}